pyproject.toml
[tool.poetry]
name = "qa-qutomation"
version = "0.1.0"
description = ""
authors = ["Danil Vorobev <hsabuse@yandex.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
pytest-xdist = "^3.3.1"
pytest-playwright = "^0.4.3"
allure-pytest = "^2.13.2"
faker = "^21.0.0"
requests = "^2.31.0"
pydantic = "^2.1.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
pymysql = "^1.1.0"
fastapi = "^0.105"
uvicorn = "^0.24"
cryptography = "^41.0.2"
black = "^23.12.0"
aiomysql = "^0.2.0"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
markers = [
    "UI",
    "API"
]
